Reworked Extensions and added README with examples
